Add reservation form and style
